refactor: mv cmd/* ./
