+++ license_bsd_clause_3.go
@@ -13,7 +13,7 @@
 
 // Parts inspired by https://github.com/ryanuber/go-license
 
-package cmd
+package main
 
 func initBsdClause3() {
 	Licenses["bsd"] = License{
