+++ license_gpl_3.go
@@ -13,7 +13,7 @@
 
 // Parts inspired by https://github.com/ryanuber/go-license
 
-package cmd
+package main
 
 func initGpl3() {
 	Licenses["gpl3"] = License{
