+++ license_mit.go
@@ -13,7 +13,7 @@
 
 // Parts inspired by https://github.com/ryanuber/go-license
 
-package cmd
+package main
 
 func initMit() {
 	Licenses["mit"] = License{
