+++ testdata/main.go.golden
@@ -15,7 +15,7 @@ limitations under the License.
 */
 package main
 
-import "github.com/spf13/cobra-cli/cmd/cmd"
+import "github.com/spf13/cobra-cli/cmd"
 
 func main() {
 	cmd.Execute()
